Ignore last bullet recoil compensation

+++ Program.cs
@@ -69,16 +69,16 @@ public class Program
                 break;
             case 7:
                 RecoilHandler.Smoothness = int.Parse(values[5]);
-                RecoilHandler.Scope = int.Parse(values[6].Replace("x", ""));
+                RecoilHandler.Scope = double.Parse(values[6].Replace("x", ""));
                 break;
             case 8:
                 RecoilHandler.Smoothness = int.Parse(values[5]);
-                RecoilHandler.Scope = int.Parse(values[6].Replace("x", ""));
+                RecoilHandler.Scope = double.Parse(values[6].Replace("x", ""));
                 RecoilHandler.GlobalOverflowCorrection = bool.Parse(values[7]);
                 break;
             case 9:
                 RecoilHandler.Smoothness = int.Parse(values[5]);
-                RecoilHandler.Scope = int.Parse(values[6].Replace("x", ""));
+                RecoilHandler.Scope = double.Parse(values[6].Replace("x", ""));
                 RecoilHandler.GlobalOverflowCorrection = bool.Parse(values[7]);
                 RecoilHandler.LocalOverflowCorrection = bool.Parse(values[8]);
                 break;
